discogs_credits: extract MB API throttle + pageWindow constant
Issue #32 — split out MB-specific code to src/api-mb.js:

  mbThrottle       the concurrency-4 cooperative-backoff throttle
                   (from the issue #30 fix)
  fetchMBEntity    /ws/js/entity/{mbid} lookup
  fetchWithRetry   thin wrapper around mbThrottle.fetchJson

Moved `pageWindow` (the `unsafeWindow ?? window` const) from the entry
top into constants.js so api-mb and future modules can share the single
derived reference.

`resolveLinkTypeId` stays in entry for now — it calls `addLogLine`,
which lives in the not-yet-extracted logger. Will move once the logger
is its own module.

Full 12-fixture suite green; no behavior change.
