+++ userscripts/discogs_credits/src/discogs_credits.user.js
@@ -9,6 +9,7 @@ import {
     REL_TEMPLATE,
     SELECTORS,
     DISCOGS_LOGO_URL,
+    pageWindow,
 }                                  from './constants.js';
 import {
     doNext,
@@ -22,6 +23,11 @@ import {
     getDiscogsLinkKey,
     getDiscogsReleaseData,
 }                                  from './api-discogs.js';
+import {
+    fetchMBEntity,
+    mbThrottle,
+    fetchWithRetry,
+}                                  from './api-mb.js';
 
 let db;
 const request = indexedDB.open('mblink');
@@ -90,8 +96,7 @@ const DISCOGS_CHANNEL = new BroadcastChannel('discogs-importer-artist');
         });
 })();
 
-// Access the real page window where MB lives.
-const pageWindow = (typeof unsafeWindow !== 'undefined') ? unsafeWindow : window;
+// `pageWindow` moved to constants.js — see import at top of this file.
 
 let lastRequest;
 let lastUiItem;
@@ -616,13 +621,6 @@ async function waitForMBEditor(timeoutMs = 15000) {
     return null;
 }
 
-/** Fetch a full MB entity from /ws/js/entity/{mbid}. */
-async function fetchMBEntity(mbid) {
-    const res = await fetch(`/ws/js/entity/${mbid}`);
-    if (!res.ok) throw new Error(`/ws/js/entity/${mbid} → ${res.status}`);
-    return res.json();
-}
-
 /**
  * Resolve a link type name to its numeric ID from MB.linkedEntities.link_type.
  *
@@ -880,90 +878,6 @@ function readIdbRecord(key) {
         } catch(e) { resolve(null); }
     });
 }
-
-// ── Centralized MB API throttle ──────────────────────────────────────────
-// All MB API requests go through this throttle. Up to MAX_CONCURRENT
-// requests can be in flight at any time (no artificial gap between
-// successive requests — MB's own backpressure paces sustained throughput).
-//
-// On 429/503 the worker that received the rate-limit pushes a shared
-// `_pauseUntil` timestamp forward by Retry-After (or an exponential backoff
-// if the header is absent). Every other worker checks `_pauseUntil` before
-// its next request and idles until it elapses — i.e. all in-flight workers
-// cooperatively back off together. This eliminates the thundering-herd
-// retry storms that the previous "5 immediate parallel threads" approach
-// produced (issue #30) while keeping the burst throughput that gave that
-// approach its speed advantage over the strict serial chain it replaced.
-const mbThrottle = (() => {
-    const MAX_CONCURRENT = 4;       // simultaneous in-flight requests
-    let _running         = 0;
-    let _pauseUntil      = 0;        // unix-ms; workers idle until this time
-    const _queue         = [];        // pending { url, retries, wantJson, resolve }
-    let _totalRequests   = 0;
-    let _rateLimited     = 0;
-
-    async function _waitForPause() {
-        let wait;
-        while ((wait = _pauseUntil - Date.now()) > 0) {
-            await new Promise(r => setTimeout(r, wait));
-        }
-    }
-
-    function _drain() {
-        while (_running < MAX_CONCURRENT && _queue.length > 0) {
-            _running++;
-            const item = _queue.shift();
-            _run(item).finally(() => { _running--; _drain(); });
-        }
-    }
-
-    async function _run(item) {
-        for (let attempt = 0; attempt <= item.retries; attempt++) {
-            await _waitForPause();
-            _totalRequests++;
-            try {
-                const res = await fetch(item.url);
-                if (res.status === 429 || res.status === 503) {
-                    _rateLimited++;
-                    const ra = parseInt(res.headers.get('Retry-After'), 10);
-                    const waitMs = (ra > 0) ? ra * 1000
-                                            : Math.min(1000 * Math.pow(2, attempt), 30000);
-                    // Push forward only — never backward — so concurrent 503s
-                    // from sibling workers don't shorten an already-pending pause.
-                    _pauseUntil = Math.max(_pauseUntil, Date.now() + waitMs);
-                    continue;
-                }
-                if (!res.ok) { item.resolve(null); return; }
-                const data = item.wantJson ? await res.json() : res;
-                item.resolve(data);
-                return;
-            } catch (e) {
-                if (attempt === item.retries) { item.resolve(null); return; }
-                await new Promise(r => setTimeout(r, 500));
-            }
-        }
-        item.resolve(null);
-    }
-
-    function _enqueue(url, retries, wantJson) {
-        return new Promise(resolve => {
-            _queue.push({ url, retries, wantJson, resolve });
-            _drain();
-        });
-    }
-
-    return {
-        fetchJson: (url, retries = 3) => _enqueue(url, retries, true),
-        fetchRaw:  (url, retries = 3) => _enqueue(url, retries, false),
-        stats: () => ({ total: _totalRequests, rateLimited: _rateLimited,
-                         inFlight: _running, queued: _queue.length }),
-    };
-})();
-
-async function fetchWithRetry(url, retries = 4) {
-    return mbThrottle.fetchJson(url, retries);
-}
-
 
 async function instantFillRelationships(companies, artistRoles, tracklistRels, applyToTracks, createWorks, discogsTracklist, processTracklist, resolvedEntityTypes, confirmedMap) {
     resolvedEntityTypes = resolvedEntityTypes || new Map();
